requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
Pillow
django-ninja==1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0